@@ -15,14 +15,12 @@
 
 """Here are the errors that can be caused by the bot."""
 
-from typing import Optional
-
 from disnake.ext.commands.errors import CommandError
 
 
 class MemberTopRolePosition(CommandError):
     """If the specified member is higher for you in the role."""
-    def __init__(self, message: Optional[str] = None) -> None:
+    def __init__(self, message: str | None = None) -> None:
         super().__init__(
             message or "the specified member above you in the role."
         )
@@ -30,7 +28,7 @@ class MemberTopRolePosition(CommandError):
 
 class MemberProtected(CommandError):
     """If the specified member is the author or owner."""
-    def __init__(self, message: Optional[str] = None) -> None:
+    def __init__(self, message: str | None = None) -> None:
         super().__init__(
             message or "the specified member is the author or owner."
         )
