@@ -38,7 +38,7 @@ async def connect_to_db(bot: Bot) -> None:
 
 
 async def close_db_connection(bot: Bot) -> None:
-    """Close the PostgreSQL database. connection."""
+    """Close the PostgreSQL database connection."""
     logger.info("Closing connection to PostgreSQL database...")
 
     try:
